Add a basic E2E test example
The test list the existing worker nodes, finding out the
IP addresses of the nodes, then it spawns 3 Pods, and inspects
the IP addresses.

Since all resources are created on individual namespace per
test item, and each namespace is destroyed when the test finishes
all the resources inside the namespace are automatically deleted
by kubernetes.

Signed-off-by: Miguel Angel Ajo <majopela@redhat.com>