Update to go mod
We use the -mod=vendor strategy for building, and we add a target
on Makefile called vendor.

Apparently go-mod expects semver (vX.Y.Z) naming for tags/branches, but
kubernetes-1.13.2 is not in that format so for k8s.io/xxx packages
it gets replaced to v0.0.0-commit id every time we run go mod,
see: https://github.com/golang/go/issues/27271

Go automatically detects ginkgo/gomega usage, and adds it to the
go.mod file, resulting in vendoring those pkgs, which may
be ok if we want to provide a built version of the e2e binary
as kubefed will be doing soon.