Submariner-engine as deployment
This modifies the deployment of the submariner engine pods to be
done as a Deployment, and we use kubernetes to handle the constraints
and placement of pods (submariner.io/gateway=true flags, etc..).

We are replicating the helm behaviour, but we may want to change the
replica count based on the size of the "Submariner" CRD handled
by the operator.