Modify the Image field, to become Repository and Version
In such way, the operator can infer the correct image names it
needs for running, the images can be local to the cluster (CI/devel),
or remote.

The default repository is 'quay.io/submariner', and the default version
now is 'latest', that would result in:

quay.io/submariner/submariner:latest
quay.io/submariner/submariner-route-agent:latest

The keyword 'local' can be used as a repository so it can be ommited
on the path to the images, for example repository=local and version=local
would result in:

submariner:local
submariner-route-agent:local

being used, and that's what we use on the e2e testing.