Commit the submariner-operator sources
To reduce the CI execution time, and avoid the issues of doing codegen
everywhere we commit the submariner-operator sources.

This is the result of checking-in the operator-sdk generated sources
via codegen-operator.

  rm -rf operators/go/submariner-operator/
  make codegen-operator
  rm operators/go/.gitignore
  git add operators/go/submariner-operator/**
  git commit

No manually generated files.

In future commits when we move this to the submariner-operator repository
we may want to remove the codegen logic, while documenting a link
for historical/reference purposes, so we avoid maintaing a codegen script
at this point.