Make globalnet detection independent of the ClusterID
We reuse the E2E testing framework for verify-connectivity, and
in that use case we don't know the cluster IDs.

This patch changes the behaviour of the framework to detect
globalnet by listing the clusters, as soon as one cluster
has Globalnet enabled it will be assumed to be true.

Related-Issue: https://github.com/submariner-io/submariner-operator/issues/381